// frontend/src/features/anamnese/types/anamnese.types.ts
export type SimNao = 'sim' | 'nao' | null;
// ...
export interface EstereotipiasRituais {
    balancaMaosLadoCorpoOuFrente: SimNao;
    balancaCorpoFrenteParaTras: SimNao;
    pulaOuGiraEmTornoDeSi: SimNao;
    repeteSonsSemFuncaoComunicativa: SimNao;
    repeteMovimentosContinuos: SimNao;
    exploraAmbienteLambendoTocando: SimNao;
    procuraObservarObjetosCantoOlho: SimNao;
    organizaObjetosLadoALado: SimNao;
    realizaTarefasSempreMesmaOrdem: SimNao;
    apresentaRituaisDiarios: SimNao;
    observacoesTopografias: string;
}

// Item 20 - Problemas de Comportamento
export interface ProblemasComportamento {
    apresentaComportamentosAutoLesivos: SimNao;
    autoLesivosQuais: string;
    apresentaComportamentosHeteroagressivos: SimNao;
    heteroagressivosQuais: string;
    apresentaDestruicaoPropriedade: SimNao;
    destruicaoDescrever: string;
    necessitouContencaoMecanica: SimNao;
    observacoesTopografias: string;
}

export interface AnamneseComportamento {
    estereotipiasRituais: EstereotipiasRituais;
    problemasComportamento: ProblemasComportamento;
}